Split home components into separate files
