+++ app/assets/scripts/views/home/home.js
@@ -1,29 +1,24 @@
-'use strict';
 import React, { useEffect, useMemo } from 'react';
 import { PropTypes as T } from 'prop-types';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import moment from 'moment';
-import * as d3 from 'd3';
 
-import { shortenLargeNumber } from '../utils/format';
 import {
   fetchRandomCompareLocs,
   fetchCompareLocationMeasurements,
   invalidateCompare,
-} from '../actions/action-creators';
-import LoadingMessage from '../components/loading-message';
-import JoinFold from '../components/join-fold';
-import Testimonials from '../components/testimonials';
-import { parameterUnit } from '../utils/map-settings';
-import { NO_CITY } from '../utils/constants';
+} from '../../actions/action-creators';
+import JoinFold from '../../components/join-fold';
+import Testimonials from '../../components/testimonials';
 
-import SponsorList from '../components/sponsor-list';
-import { getCountryName } from '../utils/countries';
-import ChartMeasurement from '../components/chart-measurement';
+import SponsorList from '../../components/sponsor-list';
 
-import sponsors from '../../content/sponsors.json';
-import testimonials from '../../content/testimonials.json';
+import sponsors from '../../../content/sponsors.json';
+import testimonials from '../../../content/testimonials.json';
+
+import CompareLocationCard from './compare-location-card';
+import StatsCount from './stats-count';
 
 function Home(props) {
   const {
@@ -270,253 +265,3 @@ function dispatcher(dispatch) {
 }
 
 module.exports = connect(selector, dispatcher)(Home);
-
-// /////////////////////////////////////////////////////////////////// //
-// Helper components
-
-function StatsCount(props) {
-  const { fetched, fetching, data } = props;
-
-  if (!fetched && !fetching) {
-    return null;
-  }
-
-  return (
-    <section className="fold fold--dark fold--type-b" id="home-fold-open-data">
-      <div className="inner">
-        <header className="fold__header">
-          <h1 className="fold__title">Open Data</h1>
-          <div className="fold__teaser prose prose--responsive">
-            <p>
-              The data is captured from multiple sources and made accessible to
-              all through our open-source platform.
-            </p>
-            <ol className="od-quick-menu">
-              <li>
-                <Link
-                  to="/locations"
-                  title="View page"
-                  className="od-quick-menu__locations-link"
-                >
-                  Location view
-                </Link>
-              </li>
-              <li>
-                <a
-                  href="https://docs.openaq.org/"
-                  target="_blank"
-                  rel="noreferrer"
-                  title="View API"
-                  className="od-quick-menu__api-link"
-                >
-                  API
-                </a>
-              </li>
-              <li>
-                <Link
-                  to="/countries"
-                  title="View page"
-                  className="od-quick-menu__countries-link"
-                >
-                  Country view
-                </Link>
-              </li>
-              <li>
-                <Link
-                  to="/compare"
-                  title="View page"
-                  className="od-quick-menu__compare-link"
-                >
-                  Compare data
-                </Link>
-              </li>
-            </ol>
-            <p className="fold__main-action">
-              <Link
-                to="/locations"
-                className="button button--large button--primary-ghost button--capsule"
-                title="View page"
-              >
-                <span>See open data</span>
-              </Link>
-              <a
-                href="https://docs.openaq.org/"
-                target="_blank"
-                rel="noreferrer"
-                className="button button--large button--white-bounded button--capsule"
-                title="View API"
-              >
-                <span>Use API</span>
-              </a>
-            </p>
-          </div>
-        </header>
-        <figure className="fold__media">
-          {!fetching ? (
-            <ol className="big-stats-list">
-              <li className="big-stat">
-                <strong className="big-stat__value">
-                  {shortenLargeNumber(data.totalMeasurements, 0)}
-                </strong>
-                <span className="big-stat__label">
-                  Air quality measurements
-                </span>
-              </li>
-              <li className="big-stat">
-                <strong className="big-stat__value">
-                  {shortenLargeNumber(data.sources, 0)}
-                </strong>
-                <span className="big-stat__label">Data sources</span>
-              </li>
-              <li className="big-stat">
-                <strong className="big-stat__value">
-                  {shortenLargeNumber(data.locations, 0)}
-                </strong>
-                <span className="big-stat__label">Locations</span>
-              </li>
-              <li className="big-stat">
-                <strong className="big-stat__value">
-                  {shortenLargeNumber(data.countries, 0)}
-                </strong>
-                <span className="big-stat__label">Countries</span>
-              </li>
-            </ol>
-          ) : (
-            <LoadingMessage />
-          )}
-        </figure>
-      </div>
-    </section>
-  );
-}
-
-StatsCount.propTypes = {
-  fetched: T.bool,
-  fetching: T.bool,
-  data: T.object,
-};
-
-function CompareLocationCard(props) {
-  const { location, measurement } = props;
-
-  if (location.error || measurement.error) {
-    return (
-      <article className="card card--measurement">
-        <div className="card__contents">
-          <div className="card__body">
-            <p>No locations with measurements were found</p>
-          </div>
-        </div>
-      </article>
-    );
-  }
-
-  if (!location.fetched || !measurement.fetched) {
-    return (
-      <article className="card card--measurement">
-        <div className="card__contents">
-          <div className="card__body">
-            <p>Loading data</p>
-          </div>
-        </div>
-      </article>
-    );
-  }
-
-  const { name, city, country, id } = location.data;
-
-  // Get recent pm25 measurement.
-  const recentMeasure = measurement.data.results[0];
-
-  return (
-    <article className="card card--measurement">
-      <Link to={`/location/${id}`} className="card__contents" title="View more">
-        <header className="card__header">
-          <div className="card__headline">
-            <h1 className="card__title">{name}</h1>
-          </div>
-        </header>
-        <div className="card__body">
-          <dl className="card--measurement__details">
-            <dt>Measurements</dt>
-            {recentMeasure && <dd>PM2.5</dd>}
-            {recentMeasure && (
-              <dd>
-                <strong>{Math.round(recentMeasure.average)}</strong>{' '}
-                <sub>{recentMeasure.unit}</sub>
-              </dd>
-            )}
-            <dt>Location</dt>
-            <dd>
-              {city || NO_CITY}
-              <span>, </span>
-              <small>{getCountryName(country) || 'N/A'}</small>
-            </dd>
-          </dl>
-          <figure className="card--measurement__chart">
-            <CompareLocationCardChart measurement={measurement} />
-          </figure>
-        </div>
-      </Link>
-    </article>
-  );
-}
-
-CompareLocationCard.propTypes = {
-  location: T.object,
-  measurement: T.object,
-};
-
-function CompareLocationCardChart(props) {
-  const { measurement } = props;
-
-  // All the times are local and shouldn't be converted to UTC.
-  // The values should be compared at the same time local to ensure an
-  // accurate comparison.
-  const dateFormat = 'YYYY/MM/DD HH:mm:ss';
-  const userNow = moment().format(dateFormat);
-  const weekAgo = moment().subtract(7, 'days').format(dateFormat);
-
-  // Prepare data.
-  const chartData = useMemo(
-    () =>
-      measurement.data.results
-        .filter(res => {
-          if (res.average < 0) return false;
-          const measurementDate = moment(res.hour).format(dateFormat);
-          return measurementDate >= weekAgo && measurementDate <= userNow;
-        })
-        .map(o => {
-          // Map data according to chart needs.
-          return {
-            value: o.average,
-            date: {
-              localNoTZ: new Date(o.hour),
-            },
-          };
-        }),
-    [measurement]
-  );
-
-  const yMax = d3.max(chartData, o => o.value) || 0;
-
-  // 1 Week.
-  const xRange = [moment().subtract(7, 'days').toDate(), moment().toDate()];
-
-  if (!chartData.length) return null;
-
-  return (
-    <ChartMeasurement
-      className="home-compare-chart"
-      data={[chartData]}
-      xRange={xRange}
-      yRange={[0, yMax]}
-      yLabel={parameterUnit['pm25']}
-      compressed
-    />
-  );
-}
-
-CompareLocationCardChart.propTypes = {
-  measurement: T.object,
-};
