+++ app/assets/scripts/views/home/index.js
@@ -1,0 +1,3 @@
+import Home from './home';
+
+export default Home;
